+++ src/CleanRegex/Match/Details/Group/ReplaceDetailGroup.php
@@ -3,6 +3,8 @@ namespace TRegx\CleanRegex\Match\Details\Group;
 
 interface ReplaceDetailGroup extends DetailGroup
 {
+    public function modifiedSubject(): string;
+
     public function modifiedOffset(): int;
 
     public function byteModifiedOffset(): int;