+++ src/CleanRegex/Match/Details/Group/ReplaceNotMatchedGroup.php
@@ -3,6 +3,11 @@ namespace TRegx\CleanRegex\Match\Details\Group;
 
 class ReplaceNotMatchedGroup extends NotMatchedGroup implements ReplaceDetailGroup, ReplaceMatchGroup
 {
+    public function modifiedSubject(): string
+    {
+        throw $this->groupNotMatched('modifiedSubject');
+    }
+
     public function modifiedOffset(): int
     {
         throw $this->groupNotMatched('modifiedOffset');