+++ src/CleanRegex/Match/Details/ReplaceDetail.php
@@ -5,11 +5,11 @@ use TRegx\CleanRegex\Match\Details\Group\ReplaceDetailGroup;
 
 interface ReplaceDetail extends Detail
 {
+    public function modifiedSubject(): string;
+
     public function modifiedOffset(): int;
 
     public function byteModifiedOffset(): int;
-
-    public function modifiedSubject(): string;
 
     public function group($nameOrIndex): ReplaceDetailGroup;
 }